upgrade to tensorflow 1.0

+++ misc/datasets.py
@@ -72,8 +72,8 @@ class Dataset(object):
                 np.zeros([images.shape[0], self._imsize, self._imsize, 3])
             ori_size = images.shape[1]
             for i in range(images.shape[0]):
-                h1 = np.floor((ori_size - self._imsize) * np.random.random())
-                w1 = np.floor((ori_size - self._imsize) * np.random.random())
+                h1 = np.floor((ori_size - self._imsize) * np.random.random()).astype(np.int)
+                w1 = np.floor((ori_size - self._imsize) * np.random.random()).astype(np.int)
                 cropped_image =\
                     images[i][w1: w1 + self._imsize, h1: h1 + self._imsize, :]
                 if random.random() > 0.5:
